requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/fmnisme/python-icinga2api.git@9a1a3cc7968d6c72bf49e97ef387b2824e6835e9#egg=icinga2api
alabaster==0.7.10         # via sphinx
ansible-lint==3.4.23      # via molecule
ansible==2.7.5
anyconfig==0.9.4          # via molecule
appdirs==1.4.3            # via openstacksdk, os-client-config
arrow==0.10.0             # via jinja2-time
asn1crypto==0.23.0        # via cryptography
babel==2.5.1              # via sphinx
bcrypt==3.1.3             # via paramiko
binaryornot==0.4.4        # via cookiecutter
cerberus==1.2             # via molecule
certifi==2017.7.27.1      # via requests
cffi==1.11.0              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via binaryornot, requests
click-completion==0.3.1   # via molecule
click==6.7                # via click-completion, cookiecutter, git-url-parse, molecule, pip-tools, python-gilt
cliff==2.14.0
cmd2==0.8.9               # via cliff
colorama==0.3.9           # via molecule, python-gilt
configparser==3.5.0       # via flake8
contextlib2==0.5.5        # via cmd2
cookiecutter==1.6.0       # via molecule
cryptography==2.4.2       # via ansible, openstacksdk, paramiko
decorator==4.1.2          # via openstacksdk
docutils==0.14            # via sphinx
dogpile.cache==0.6.4      # via openstacksdk
enum34==1.1.6             # via cmd2, cryptography, flake8
fasteners==0.14.1         # via python-gilt
flake8==3.5.0             # via molecule
future==0.16.0            # via cookiecutter
futures==3.1.1            # via openstacksdk
git-url-parse==1.0.1      # via python-gilt
idna==2.6                 # via cryptography, requests
imagesize==0.7.1          # via sphinx
ipaddress==1.0.18         # via cryptography, openstacksdk
iso8601==0.1.12           # via keystoneauth1, openstacksdk
jinja2-time==0.2.0        # via cookiecutter
jinja2==2.10              # via ansible, click-completion, cookiecutter, jinja2-time, molecule, sphinx
jmespath==0.9.3           # via openstacksdk
jsonpatch==1.16           # via openstacksdk
jsonpointer==1.12         # via jsonpatch
keystoneauth1==3.11.2     # via openstacksdk, os-client-config
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8
molecule==2.16.0
monotonic==1.3            # via fasteners
munch==2.2.0              # via openstacksdk
netaddr==0.7.19
netifaces==0.10.6         # via openstacksdk
openstacksdk==0.22.0
os-client-config==1.28.0  # via shade
os-service-types==1.4.0   # via keystoneauth1, openstacksdk
packaging==18.0           # via sphinx
paramiko==2.3.1           # via ansible
pathspec==0.5.5           # via yamllint
pbr==4.0.4                # via cliff, git-url-parse, keystoneauth1, molecule, openstacksdk, os-service-types, python-gilt, shade, stevedore
pexpect==4.6.0            # via molecule
pip-tools==3.2.0
poyo==0.4.1               # via cookiecutter
prettytable==0.7.2        # via cliff
psutil==5.4.6             # via molecule
ptyprocess==0.5.2         # via pexpect
py==1.4.34                # via pytest
pyasn1==0.3.6             # via paramiko
pycodestyle==2.3.1        # via flake8
pycparser==2.18           # via cffi
pyflakes==1.5.0           # via flake8
pygments==2.2.0           # via sphinx
pynacl==1.1.2             # via paramiko
pyparsing==2.3.0          # via cliff, cmd2, packaging
pyperclip==1.7.0          # via cmd2
pytest==3.2.2             # via testinfra
python-dateutil==2.6.1    # via arrow
python-gilt==1.2.1        # via molecule
pytz==2017.2              # via babel
pyyaml==3.12              # via ansible, ansible-lint, cliff, molecule, openstacksdk, os-client-config, python-gilt, yamllint
requests==2.18.4          # via cookiecutter, keystoneauth1, sphinx
requestsexceptions==1.3.0  # via openstacksdk, os-client-config
sh==1.12.14               # via molecule, python-gilt
shade==1.30.0
six==1.11.0               # via ansible-lint, bcrypt, click-completion, cliff, cmd2, cryptography, fasteners, git-url-parse, keystoneauth1, molecule, munch, openstacksdk, packaging, pip-tools, pynacl, python-dateutil, sphinx, stevedore, testinfra
snowballstemmer==1.2.1    # via sphinx
sphinx==1.8.3
sphinxcontrib-websupport==1.0.1  # via sphinx
stevedore==1.26.0         # via cliff, keystoneauth1
subprocess32==3.5.3       # via cmd2
tabulate==0.8.2           # via molecule
testinfra==1.14.0
tree-format==0.1.2        # via molecule
typing==3.6.2             # via sphinx
unicodecsv==0.14.1        # via cliff
urllib3==1.22             # via requests
wcwidth==0.1.7            # via cmd2
whichcraft==0.4.1         # via cookiecutter
yamllint==1.11.1          # via molecule
